+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [11.19.1](https://github.com/sweetalert2/sweetalert2/compare/v11.19.0...v11.19.1) (2025-04-15)
+
+
+### Bug Fixes
+
+* shift+tab when a popup is focused ([#2815](https://github.com/sweetalert2/sweetalert2/issues/2815)) ([f68b60e](https://github.com/sweetalert2/sweetalert2/commit/f68b60e9fefd713b2b50ec15b94d5577fdced1ae))
+
+# [11.19.0](https://github.com/sweetalert2/sweetalert2/compare/v11.18.0...v11.19.0) (2025-04-15)
+
+
+### Features
+
+* match outline color with custom button colors ([#2814](https://github.com/sweetalert2/sweetalert2/issues/2814)) ([62fb3fc](https://github.com/sweetalert2/sweetalert2/commit/62fb3fc8eaaacf5931a5d541003f479cb992f6d0))
+
+# [11.18.0](https://github.com/sweetalert2/sweetalert2/compare/v11.17.2...v11.18.0) (2025-04-14)
+
+
+### Features
+
+* add embed-iframe theme ([#2813](https://github.com/sweetalert2/sweetalert2/issues/2813)) ([f455b5f](https://github.com/sweetalert2/sweetalert2/commit/f455b5f93291f07c30fd1557f0a35497ea6b7592))
+
 ## [11.17.2](https://github.com/sweetalert2/sweetalert2/compare/v11.17.1...v11.17.2) (2025-02-16)